inline fbpatch.py (#232)
* inline fbpatch.py

* use constant instead of magic number

* use Vertex3 methods
