+++ piqueserver/config/config.json
@@ -132,7 +132,6 @@
         "piqueserver.scripts.ratio",
         "piqueserver.scripts.passreload",
         "piqueserver.scripts.blockinfo",
-        "piqueserver.scripts.fbpatch",
         "piqueserver.scripts.afk"
     ],
