+++ piqueserver/player.py
@@ -1,14 +1,17 @@
 from __future__ import print_function, unicode_literals
 
+import math
+
 from twisted.internet import reactor
 
 from piqueserver import commands
 
 import pyspades
 from pyspades.constants import (ERROR_BANNED, DESTROY_BLOCK, SPADE_DESTROY,
-                                GRENADE_DESTROY, ERROR_KICKED)
+                                GRENADE_DESTROY, ERROR_KICKED, BLOCK_TOOL)
 from pyspades.server import ServerConnection
-from pyspades.common import encode, prettify_timespan
+from pyspades.common import encode, prettify_timespan, Vertex3
+from pyspades.world import Character
 
 # TODO: move these where they belong
 CHAT_WINDOW_SIZE = 5
@@ -34,6 +37,7 @@ class FeatureConnection(ServerConnection):
     chat_count = 0
     user_types = None
     rights = None
+    can_complete_line_build = True
 
     def on_connect(self):
         protocol = self.protocol
@@ -113,6 +117,8 @@ class FeatureConnection(ServerConnection):
         print(log_message.replace("\n", "\\n").encode('ascii', 'replace'))
 
     def _can_build(self):
+        if not self.can_complete_line_build:
+            return False
         if not self.building:
             return False
         if not self.god and not self.protocol.building:
@@ -120,6 +126,37 @@ class FeatureConnection(ServerConnection):
 
     def on_block_build_attempt(self, x, y, z):
         return self._can_build()
+
+    def on_secondary_fire_set(self, secondary):
+
+        # Inlined from fbpatch.py
+        # Author: Nick Christensen AKA a_girl
+        # Distant Drag Build Client Bug Patch for (0.75) and possibly (0.76)
+        #
+        # if right mouse button has been clicked to initiate drag building;
+        # distinguishes from the right click release that marks the end point.
+        if secondary:
+            if self.tool == BLOCK_TOOL:  # 1 refers to block tool; if the tool in hand is a block
+                # grab player current position at drag build start
+                position = self.world_object.position
+                # grab player current orientation at drag build start
+                vector = self.world_object.orientation
+                # probably unnecessary, but makes sure vector values are
+                # between 0 and 1 inclusive
+                vector.normalize()
+                # creates a line object starting at player and following
+                # their point of view.
+                c = Character(self.world_object.world, position, vector)
+                # finds coordinates of the first block this line strikes.
+                line_start = c.cast_ray()
+                if line_start:  # if player is pointing at a valid point.  Distant solid blocks will return False
+                    distance = (Vertex3(*line_start) - Vertex3(position.x, position.y, position.z)).length()
+                    if distance > 6:
+                        self.can_complete_line_build = False
+                    else:
+                        self.can_complete_line_build = True
+                else:
+                    self.can_complete_line_build = False
 
     def on_line_build_attempt(self, points):
         if self._can_build() == False:
